+++ plugin/src/net/groboclown/idea/p4ic/v2/server/connection/ProjectConfigSourceLoader.java
@@ -59,9 +59,12 @@ public class ProjectConfigSourceLoader {
 
             // Relative config file.  Make sure that we use the version
             // that maps the correct root directory to the config file.
+            LOG.debug("Loading p4config files named " + configFile + " under " +
+                    P4Vcs.getInstance(project).getVcsRoots());
             Map<VirtualFile, P4Config> map = P4ConfigUtil.loadCorrectDirectoryP4Configs(project, configFile);
             if (map.isEmpty()) {
-                LOG.info("Config invalid because no p4config files were found");
+                LOG.info("Config invalid because no p4config files were found.  Searching for config files named " +
+                    configFile + ", under vcs roots " + P4Vcs.getInstance(project).getVcsRoots());
                 final Builder builder = new Builder(project, config);
                 builder.setError(new P4InvalidConfigException(P4Bundle.message("error.config.no-file")));
                 return Collections.singleton(builder);